requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
distlib==0.3.6
Django==3.0
django-advanced-filters==2.0.0
django-filter==1.1.0
django-templated-mail==1.1.1
django-utils==0.0.2
django-utils-six==2.0
djangorestframework==3.12.4
djangorestframework-filters==0.11.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
filelock==3.8.0
idna==3.4
importlib-metadata==1.7.0
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
latest==0.6.0
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
pipenv==2022.10.25
platformdirs==2.5.2
pluggy==0.13.1
py==1.11.0
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.1
simplejson==3.17.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.6
virtualenv-clone==0.5.7
zipp==3.10.0
